requirements.txt
flask==2.3.2
gunicorn==20.1.0
praw==7.7.1
python-dotenv==1.0.0
openai==0.28.0
Flask-Limiter==3.5.0
requests==2.31.0
